+++ gitolith-core/src/changelog/mod.rs
@@ -1,0 +1,4 @@
+/// Template context.
+pub mod context;
+/// Changelog generator.
+pub mod generator;