+++ templates/CHANGELOG.md
@@ -1,1 +1,8 @@
 ## {{ release_title }}
+
+{% for change in changes %}
+### {{ change.title }}
+{% for entry in change.entries %}
+- {{ entry }}
+{% endfor %}
+{% endfor %}